refactor: move out the now nicely generic text utils
